pyproject.toml
[project]
name = "kubernetes-docs-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "bs4>=0.0.2",
    "html2text>=2024.2.26",
    "readability-lxml>=0.8.1",
    "requests>=2.32.3",
    "tqdm>=4.67.1",
]
